scope home sections to components
