+++ components/ui/big-card/styles.scss
@@ -1,0 +1,44 @@
+@import '~styles/settings.scss';
+
+.c-big-card {
+  position: relative;
+  background-color: $grey-light;
+  border: solid 1px $border;
+  overflow: hidden;
+  padding: 30px;
+
+  @media screen and (min-width: $screen-m) {
+    padding: 40px;
+  }
+
+  .big-card-content {
+    position: relative;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    z-index: 1;
+
+    h3,
+    p {
+      margin-bottom: 15px;
+    }
+
+    h3 {
+      font-size: 30px;
+    }
+
+    p {
+      color: $slate;
+    }
+  }
+
+  .big-card-image {
+    position: absolute;
+    bottom: -150px;
+    left: -100px;
+    width: 120%;
+    background-size: contain;
+    z-index: 0;
+  }
+}
