+++ layouts/home/component.jsx
@@ -1,129 +1,28 @@
-import React, { useState, useRef } from 'react';
 import PropTypes from 'prop-types';
-import YouTube from 'react-youtube';
 import Link from 'next/link';
 import cx from 'classnames';
 
 import { Desktop, Mobile, Carousel, Button, Row, Column } from 'gfw-components';
 
-import Cover from 'components/cover';
 import Icon from 'components/ui/icon';
 import Card from 'components/ui/card';
 import NoContent from 'components/ui/no-content';
 
-import arrowIcon from 'assets/icons/arrow-down.svg?sprite';
 import profileIcon from 'assets/icons/profile.svg?sprite';
-import mailIcon from 'assets/icons/mail.svg?sprite';
 
 import config from './config';
 import newsImage from './assets/news-bg.jpg';
-import bgImage from './assets/home-bg.jpg';
+
+import HomeCover from './cover';
+import HomeSummary from './summary';
 
 import './styles.scss';
 
-const HomePage = ({ summary, uses, apps, news }) => {
-  const [showVideo, setShowVideo] = useState(false);
-  const summaryEl = useRef(null);
-
+const HomePage = ({ uses, apps, news }) => {
   return (
     <div className="l-home-page">
-      <Cover
-        className="section-cover"
-        title="Forest Monitoring Designed for Action"
-        description="Global Forest Watch offers the latest data, technology and tools that empower people everywhere to better protect forests."
-        bgImage={bgImage}
-        bgAlt="View of the earth from space"
-        large
-      >
-        <>
-          <div className={cx('home-video', { '-show': showVideo })}>
-            <YouTube
-              videoId="0XsJNU75Si0"
-              opts={{
-                height: '100%',
-                width: '100%',
-                playerVars: {
-                  autoplay: 1,
-                  autohide: 1,
-                  loop: 1,
-                  modestbranding: 1,
-                  rel: 0,
-                  showinfo: 0,
-                  controls: 0,
-                  disablekb: 1,
-                  enablejsapi: 0,
-                  iv_load_policy: 3,
-                },
-              }}
-              onPlay={() => setTimeout(() => setShowVideo(true), 300)}
-              onEnd={() => setShowVideo(false)}
-            />
-          </div>
-          {showVideo && (
-            <Button
-              className="stop-video-btn"
-              onClick={() => setShowVideo(false)}
-            >
-              STOP VIDEO
-            </Button>
-          )}
-          <Link href="/subscribe">
-            <a className="subscribe-link">
-              <Button round className="subscribe-btn">
-                <Icon icon={mailIcon} />
-              </Button>
-              <span className="subscribe-msg">
-                SUBSCRIBE TO THE GFW NEWSLETTER
-              </span>
-            </a>
-          </Link>
-        </>
-      </Cover>
-      <Row>
-        <Column>
-          <div className="section-summary" ref={summaryEl}>
-            <Button
-              className="scroll-to-btn"
-              round
-              onClick={() => {
-                window.scrollTo({
-                  behavior: 'smooth',
-                  left: 0,
-                  top: summaryEl?.current?.offsetTop,
-                });
-              }}
-            >
-              <Icon icon={arrowIcon} />
-            </Button>
-            {summary && (
-              <>
-                <Desktop>
-                  <Carousel settings={{ dots: true }}>
-                    {summary.map((c) => (
-                      <Card
-                        className="summary-card"
-                        key={c.title}
-                        data={{ ...c, fullSummary: true }}
-                      />
-                    ))}
-                  </Carousel>
-                </Desktop>
-                <Mobile>
-                  <Carousel settings={{ dots: true, slidesToShow: 1 }}>
-                    {summary.map((c) => (
-                      <Card
-                        className="summary-card"
-                        key={c.title}
-                        data={{ ...c, fullSummary: true }}
-                      />
-                    ))}
-                  </Carousel>
-                </Mobile>
-              </>
-            )}
-          </div>
-        </Column>
-      </Row>
+      <HomeCover />
+      <HomeSummary />
       <div className="section-uses">
         <h3 className="section-title">
           What can you do with Global Forest Watch?
@@ -302,7 +201,6 @@ const HomePage = ({ summary, uses, apps, news }) => {
 };
 
 HomePage.propTypes = {
-  summary: PropTypes.array.isRequired,
   apps: PropTypes.array.isRequired,
   news: PropTypes.array.isRequired,
   uses: PropTypes.array.isRequired,
