+++ layouts/home/cover/index.js
@@ -1,0 +1,3 @@
+import Component from './component';
+
+export default Component;
