+++ styles/base.scss
@@ -74,7 +74,6 @@ a {
 
 a {
   color: $green-gfw;
-  border: none;
 
   &:hover {
     color: $dark-green;
